TI814X: Fixes the i2c clock issue.
The I2C and UART clock is derived from sysclk10
using DIV B using an input clock freq of 192MHz.
The ROM code sets the DIV to 4 to generate a 48MHz
for the UART while I2C PISYSCLK needs 96 MHz.

This patch fixes this conflict by setting the
DIV value to 2 to generate a clk of 96MHz and the
UART driver is configured to take 96MHz as input
in calculating the baud rate divisor.

Signed-off-by: Raj, Deepu <deepu.raj@ti.com>