TI81XX: Enable support for D-cache and MMU for TI81XX
This patch enables the supprot for D-cache with MMU
identical mapping. I-cache is enabled by default in
the start.S but D-cache is not enable by default and it
can be manage by the set of following apis.
	dcache_enable()
	dcache_disable()
	invalidate_dcache_range()
	flush_dcache_rage()

- D-cache operations supported:
        - Invalidate entire D-cache
        - Invalidate D-cache range
        - Flush(clean & invalidate) entire D-cache
        - Flush D-cache range

This patch is basesd on the following patches for MMU
support and cache maintance for armv7

http://git.denx.de/?p=u-boot.git;a=commitdiff;h=880eff5cfb9df6f0855f4e48affd349ca64692e9
        ARM: cp15: setup mmu and enable dcache

http://git.denx.de/?p=u-boot.git;a=commitdiff;h=2c451f7831208741d0ff7ca6046cffcd9ee49def
        Add a framework for layered cache maintenance
        separate out SOC specific outer cache maintenance from
        maintenance of caches known to CPU

Signed-off-by: Raj, Deepu <deepu.raj@ti.com>