AM335x : Add USB support for AM335x in u-boot
Adds USB support in uboot for AM335x.
The support for USB0 and USB1 can be added dynamically
by changing the boards.cfg file for USB0 or USB1.
This patch adds support for USB0.
USB 1 has a full size connector so acts in host mode and
USB 0 has a mini connector so used in device mode.

Signed-off-by: Gene Zarkhin <gene_zarkhin@bose.com>
Signed-off-by: Harman Sohanpal <harman_sohanpal@ti.com>